+++ apps/dashboard/src/app/account/account-users/account-users.component.ts
@@ -1,10 +1,10 @@
 import { ChangeDetectionStrategy, ChangeDetectorRef, Component, OnDestroy, OnInit } from '@angular/core';
 import { UntypedFormControl, Validators } from '@angular/forms';
 import { TranslateService } from '@ngx-translate/core';
-import { Observable, of, Subject } from 'rxjs';
-import { filter, map, switchMap, takeUntil, tap } from 'rxjs/operators';
+import { forkJoin, Observable, of, Subject, take } from 'rxjs';
+import { catchError, filter, map, switchMap, takeUntil, tap } from 'rxjs/operators';
 import { SDKService } from '@flaps/core';
-import { Account, AccountRoles, AccountUsersPayload, FullAccountUser } from '@nuclia/core';
+import { Account, AccountRoles, AccountUsersPayload, FullAccountUser, PendingInvitation } from '@nuclia/core';
 import { SisModalService, SisToastService } from '@nuclia/sistema';
 
 @Component({
@@ -15,7 +15,8 @@ import { SisModalService, SisToastService } from '@nuclia/sistema';
 })
 export class AccountUsersComponent implements OnDestroy, OnInit {
   account?: Account;
-  users?: FullAccountUser[];
+  users: FullAccountUser[] = [];
+  invitations: PendingInvitation[] = [];
   email = new UntypedFormControl([''], [Validators.required, Validators.email]);
 
   roleTranslations: { [role: string]: string } = {
@@ -42,23 +43,27 @@ export class AccountUsersComponent implements OnDestroy, OnInit {
     this.account$
       .pipe(
         tap((account) => {
-          this.account = account!;
+          this.account = account;
         }),
-        switchMap(() => this.updateUsers()),
+        switchMap((account) => this.updateUsers(account)),
         takeUntil(this.unsubscribeAll),
       )
-      .subscribe(() => this.cdr?.markForCheck());
+      .subscribe(() => this.cdr.markForCheck());
   }
 
-  isItMe(userId: string) {
+  isItMe(userId?: string) {
     return this.sdk.nuclia.auth.getJWTUser()?.sub === userId;
   }
 
-  updateUsers(): Observable<FullAccountUser[]> {
-    return this.sdk.nuclia.db.getAccountUsers(this.account!.slug).pipe(
-      tap((users) => {
+  updateUsers(account: Account): Observable<void> {
+    return forkJoin([
+      this.sdk.nuclia.db.getAccountUsers(account.slug),
+      this.sdk.nuclia.db.getAccountInvitations(account.id).pipe(catchError(() => of([]))),
+    ]).pipe(
+      map(([users, pendingInvitations]) => {
         this.users = users;
-        this.cdr?.markForCheck();
+        this.invitations = pendingInvitations;
+        this.cdr.markForCheck();
       }),
     );
   }
@@ -68,14 +73,17 @@ export class AccountUsersComponent implements OnDestroy, OnInit {
     this.sdk.nuclia.db.inviteToAccount(this.account!.slug, data).subscribe(() => {
       this.toaster.success(this.translate.instant('account.invited_user', { user: this.email.value }));
       this.email.patchValue('');
-      this.cdr?.markForCheck();
+      this.cdr.markForCheck();
     });
   }
 
   changeRole(user: FullAccountUser, role: string): void {
     if (role === 'AMEMBER') {
       this._changeRole(user, role)
-        .pipe(switchMap(() => this.updateUsers()))
+        .pipe(
+          switchMap(() => this.account$.pipe(take(1))),
+          switchMap((account) => this.updateUsers(account)),
+        )
         .subscribe();
     } else {
       this.modalService
@@ -85,7 +93,8 @@ export class AccountUsersComponent implements OnDestroy, OnInit {
         })
         .onClose.pipe(
           switchMap((result) => (!!result ? this._changeRole(user, role as AccountRoles) : of(null))),
-          switchMap(() => this.updateUsers()),
+          switchMap(() => this.account$.pipe(take(1))),
+          switchMap((account) => this.updateUsers(account)),
           takeUntil(this.unsubscribeAll),
         )
         .subscribe();
@@ -107,9 +116,10 @@ export class AccountUsersComponent implements OnDestroy, OnInit {
         ),
         filter((confirm) => !!confirm),
         switchMap(() => this.deleteUser(user)),
-        switchMap(() => this.updateUsers()),
-        switchMap(() => this.sdk.nuclia.db.getAccount(this.account!.slug)),
-        takeUntil(this.unsubscribeAll),
+        switchMap(() => this.account$.pipe(take(1))),
+        switchMap((account) =>
+          this.updateUsers(account).pipe(switchMap(() => this.sdk.nuclia.db.getAccount(account.slug))),
+        ),
       )
       .subscribe((account) => (this.sdk.account = account));
   }
@@ -126,6 +136,24 @@ export class AccountUsersComponent implements OnDestroy, OnInit {
       delete: [user.id],
     };
     return this.sdk.nuclia.db.setAccountUsers(this.account!.slug, users);
+  }
+
+  deleteInvitation(invite: PendingInvitation) {
+    this.account$
+      .pipe(
+        take(1),
+        switchMap((account) => this.sdk.nuclia.db.deleteAccountInvitation(account.id, invite.email)),
+        tap(() => this.toaster.success(`${invite.email} invitation deleted.`)),
+        catchError(() => {
+          this.toaster.error(`An error occurred while deleting ${invite.email} invitation.`);
+          return of();
+        }),
+        switchMap(() => this.account$.pipe(take(1))),
+        switchMap((account) =>
+          this.updateUsers(account).pipe(switchMap(() => this.sdk.nuclia.db.getAccount(account.slug))),
+        ),
+      )
+      .subscribe((account) => (this.sdk.account = account));
   }
 
   ngOnDestroy() {