+++ libs/common/src/lib/validators/form.validator.ts
@@ -1,10 +1,10 @@
-import { AbstractControl, ValidationErrors, ValidatorFn } from '@angular/forms';
+import { AbstractControl, ValidatorFn } from '@angular/forms';
 
 const forbiddenSlugPattern = new RegExp(/[^\w-_]+/g);
 export function Sluggable(kbSlug = false): ValidatorFn {
   return (control: AbstractControl) => {
     const forbidden = forbiddenSlugPattern.test(control.value);
-    return forbidden ? { sluggable: kbSlug ? 'kb.invalid-slug' : 'validation.invalid-slug' } : null;
+    return forbidden ? { sluggable: kbSlug ? 'kb.form.invalid-slug' : 'validation.invalid-slug' } : null;
   };
 }