+++ libs/core/src/lib/api/zone.service.ts
@@ -1,12 +1,12 @@
 import { Injectable } from '@angular/core';
 import { Zone } from '../models';
-import { forkJoin, map, Observable, switchMap } from 'rxjs';
+import { forkJoin, map, Observable, of, switchMap } from 'rxjs';
 import { FeatureFlagService } from '../analytics/feature-flag.service';
 import { SDKService } from './sdk.service';
 import { UserService } from './user.service';
 import { take } from 'rxjs/operators';
+import { BillingService } from './billing.service';
 
-const VERSION = 'v1';
 const ZONES = 'zones';
 
 @Injectable({
@@ -17,6 +17,7 @@ export class ZoneService {
     private sdk: SDKService,
     private featureFlagService: FeatureFlagService,
     private userService: UserService,
+    private billingService: BillingService,
   ) {}
 
   getZones(includeZonesBlocked = false): Observable<Zone[]> {
@@ -39,6 +40,21 @@ export class ZoneService {
           }),
         );
       }),
+      switchMap((zones) =>
+        // Keep only AWS zones on AWS_MARKETPLACE subscriptions
+        // (always keeps all zones on stage as there is only one zone there)
+        this.featureFlagService.isStageOrDev
+          ? of(zones)
+          : this.billingService
+              .getSubscription()
+              .pipe(
+                map((subscription) =>
+                  subscription && subscription.provider === 'AWS_MARKETPLACE'
+                    ? zones.filter((zone) => zone.cloud_provider === 'AWS')
+                    : zones,
+                ),
+              ),
+      ),
     );
   }
 }